backend_tasks/delivery-server/pyproject.toml
[tool.poetry]
name = "delivery"
version = "0.1.0"
description = ""
authors = ["Nguyễn Danh Trung <danhtrung0307@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
redis = "^5.0.4"
boto3 = "^1.34.99"
django-model-utils = "4.2.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
